backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.6.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
Pillow==10.2.0
openai==1.10.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
requests==2.31.0
